moevd new util assert to utils file
